Parallel: fix trie prefetch on dispatcher
Trie prefetch will be scheduled on object finalize.
With parallel, we should schedule trie prefetch on dispatcher, since
the TriePrefetcher is not safe for concurrent access and it is created & stopped
on dispatcher routine.

But object.finalize on slot cleared its dirtyStorage, which broken the later trie
prefetch on dispatcher when do MergeSlotDB.